refactor: relocation main module logic, - migrate to calendar logic with module - migrate to sendErrorEmail with notifer module.
